+++ app/layout.js
@@ -1,7 +1,7 @@
 import "@styles/globals.css";
 import { Inter } from "next/font/google";
 import Head from "next/head";
-import Sidebar from "@components/Sidebar";
+import Sidebar from "@components/Sidebar/Sidebar";
 
 const inter = Inter({ subsets: ["latin"] });
